wip: inicial migration

+++ supabase/migrations/20240101000000_initial_schema.sql
@@ -214,7 +214,7 @@ BEGIN
     ) VALUES (
         v_company_id,
         payload->>'address',
-        payload->>'number',
+        (payload->>'number')::integer,
         payload->>'complement',
         payload->>'city',
         payload->>'state',
